### 2025-06-12 #8397 @editorialbot
[PRE REVIEW]: MCSimMod: An R Package for Working with Ordinary Differential Equation Models Encoded in the MCSim Model Specification Language
**Submitting author:** @dfkapraun (<a href="http://orcid.org/0000-0001-5570-6383">Dustin F. Kapraun</a>)
**Repository:** https://github.com/USEPA/MCSimMod
**Branch with paper.md** (empty if default branch): manuscript
**Version:** v0.9.1.9000
