### 2025-06-12 #8397 @danielskatz
👋 @dfkapraun - thanks for your submission.

I don't see a license file in your repository. Can you discuss this briefly?

In addition, you could work on the possibly missing DOIs that editorialbot suggests, but note that some may be incorrect. Please feel free to make changes to your .bib file, then use the command `@editorialbot check references` to check again, and the command `@editorialbot generate pdf` when the references are right to make a new PDF. editorialbot commands need to be the first entry in a new comment.
