+++ paper/paper.bib
@@ -1,47 +1,48 @@
 @manual{R,
-    title = {R: A Language and Environment for Statistical Computing},
-    author = {{R Core Team}},
-    organization = {R Foundation for Statistical Computing},
-    address = {Vienna, Austria},
-    year = {2024},
-    url = {https://www.R-project.org/},
-  }
+  title = {R: A Language and Environment for Statistical Computing},
+  author = {{R Core Team}},
+  organization = {R Foundation for Statistical Computing},
+  address = {Vienna, Austria},
+  year = {2024},
+  url = {https://www.R-project.org/},
+}
 
 @article{deSolve,
-    title = {Solving Differential Equations in {R}: Package de{S}olve},
-    author = {Karline Soetaert and Thomas Petzoldt and R. Woodrow Setzer},
-    journal = {Journal of Statistical Software},
-    volume = {33},
-    number = {9},
-    pages = {1--25},
-    year = {2010},
-    doi = {10.18637/jss.v033.i09},
-  }
+  title = {Solving Differential Equations in {R}: Package de{S}olve},
+  author = {Karline Soetaert and Thomas Petzoldt and R. Woodrow Setzer},
+  journal = {Journal of Statistical Software},
+  volume = {33},
+  number = {9},
+  pages = {1--25},
+  year = {2010},
+  doi = {10.18637/jss.v033.i09},
+}
 
 @manual{deSolve_documentation,
-    title = {deSolve: Solvers for Initial Value Problems of Differential Equations},
-    author = {Karline Soetaert and Thomas Petzoldt and R. Woodrow Setzer and Peter N. Brown and George D. Byrne and Ernst Hairer and Alan C. Hindmarsh and Cleve Moler and Linda R. Petzold and Youcef Saad and Clement W. Ulrich},
-    year = {2023},
-    note = {R package version 1.40},
-    url = {https://cran.r-project.org/web/packages/deSolve/deSolve.pdf},
+  title = {deSolve: Solvers for Initial Value Problems of Differential Equations},
+  author = {Karline Soetaert and Thomas Petzoldt and R. Woodrow Setzer and Peter N. Brown and George D. Byrne and Ernst Hairer and Alan C. Hindmarsh and Cleve Moler and Linda R. Petzold and Youcef Saad and Clement W. Ulrich},
+  year = {2023},
+  note = {R package version 1.40},
+  url = {https://cran.r-project.org/web/packages/deSolve/deSolve.pdf},
 }
 
 @manual{mrgsolve,
-    title = {mrgsolve: Simulate from ODE-Based Models},
-    author = {Kyle T Baron},
-    year = {2024},
-    note = {R package version 1.5.2},
-    url = {https://CRAN.R-project.org/package=mrgsolve},
-  }
+  title = {mrgsolve: Simulate from ODE-Based Models},
+  author = {Kyle T Baron},
+  year = {2024},
+  note = {R package version 1.5.2},
+  url = {https://CRAN.R-project.org/package=mrgsolve},
+  doi = {10.32614/cran.package.mrgsolve},
+}
 
 @article{MCSim,
-   title = {GNU MCSim: Bayesian Statistical Inference for SBML-Coded Systems Biology Models},
-   author = {Frederic Y. Bois},
-   journal = {Bioinformatics},
-   volume = {25},
-   pages = {1453--1454},
-   year = {2009},
-   doi = {10.1093/bioinformatics/btp162},
+  title = {GNU MCSim: Bayesian Statistical Inference for SBML-Coded Systems Biology Models},
+  author = {Frederic Y. Bois},
+  journal = {Bioinformatics},
+  volume = {25},
[…]
