### 2025-06-16 #8397 @dfkapraun
@danielskatz - Ok, thanks for the update!

On Mon, Jun 16, 2025 at 2:35 PM Daniel S. Katz ***@***.***>
wrote:

> *danielskatz* left a comment (openjournals/joss-reviews#8397)
> <https://github.com/openjournals/joss-reviews/issues/8397#issuecomment-2977658611>
>
> @dfkapraun <https://github.com/dfkapraun> - in any case, this seems ready
> to get started. However, I don't have an editor for it currently, so I'm
> going to put it on hold until an editor becomes available.
>
