### 2025-06-23 #8397 @dfkapraun
@danielskatz :

Thank you for serving as editor for this manuscript!

I sent a list of potential reviewers earlier in this thread, but will copy
the list here. The first four are researchers that I know that have
relevant expertise and that I believe will be interested in the R package
described in our manuscript. The last two are people from the JOSS list
that seem to have relevant expertise.
- Zhoumeng Lin, University of Florida, ***@***.***
- Matt Linakis, Ramboll, ***@***.***
- Nanhung Hsieh, California Department of Pesticide Regulation,
***@***.***
- Xiaoqing Chang, Inotiv, ***@***.***
- Amir Shahmoradi, University of Texas, ***@***.*** (from the
Open Journal "list of reviewers")
- elauksap (from the Open Journal "list of reviewers" )

Thanks again!

Best regards,
Dustin Kapraun

On Mon, Jun 23, 2025 at 4:20 PM Daniel S. Katz ***@***.***>
wrote:

> *danielskatz* left a comment (openjournals/joss-reviews#8397)
> <https://github.com/openjournals/joss-reviews/issues/8397#issuecomment-2997830260>
>
> 👋 @dfkapraun <https://github.com/dfkapraun> - as you can see, I'll edit
> your submission.
>
> Please suggest ~5 potential reviewers. You can search the list of people
> <https://reviewers.joss.theoj.org/lookup> that have already agreed to
> review and may be suitable for this submission. Or people who aren't in the
> JOSS system would also be ok. If you know their GitHub usernames, then
> please mention them here in this thread (without tagging them with an @).
>
