### 2025-07-01 #8397 @shahmoradi
Thank you for your invitation. I apologize for my delayed response. Unfortunately, I have been traveling over the past 10 days and am still without stable internet and the necessary tools for an effective and timely review.
Sincerely,
