### 2025-07-18 #8492 @nanhung
Hello,

Many thanks to the entire `MCSimMod` developing team for providing such a useful package. I believe it will be helpful for modelers, especially in areas of Tox/Pharma, who need a feasible tool for PBPK modeling.

I have several years using **MCSim** as my main modeling tools in the PBPK research, and **R** for subsequent statistical analysis. But usually feel annoying to exchange the data between **R** and **MCSim**. This "all-in-one" concept is an ideal solution, offering modelers like myself significant relief in coupling these headache.

After reviewing the package, documentation, and software paper.I'm genuinely impressed by how well `MCSimMod` integrates **MCSim** and **R**. I found the approach is very similair to `rstan`, the well-known package for Bayesian modeling, and I find this user-friendly design truly remarkable.

For package and documentation:

I find this software package quite practical in its design, but its functionality is a bit sketchy. This is because its main purpose is merely to allow modeling using MCSim syntax and then compile it for dynamic simulations within the `deSolve` package. Its simulation capabilities are limited to straightforward deterministic simulations, unfortunately neglects `MCSim`'s most powerful feature: probabilistic simulation (Monte Carlo and Markov Chain Monte Carlo).

I hope the developers can integrate probabilistic simulation functionality into `MCSimMod`, or at least include alternative feasible methods in the vignette. This would greatly assist users who wish to delve deeper into simulation results, making the package more accessible for advanced research.

For Software Paper:

The summary and statement of need are well-written, with the crucial literature cited. However, line 42 contains an unclear description: 'this system encourages modelers to use separate files for defining models…' Please clarify what kind of separate files are being referred to here. In addition, I suggest adding the technical detail or example of use to the manuscript. This might be helpful for readers to gain more insight into this package. For reference, the `ssdtools` package (https://joss.theoj.org/papers/10.21105/joss.07492), which I previously reviewed, offers a good example."

The other suggested improvement is to analyze the computational efficiency of `MCSimMod` relative to standalone **MCSim** or alternative software packages, such as `mrgsolve`. In my experience, the standalone MCSim has better efficiency compared with the `deSolve` (using compiled C language). But just my personal curiosity.

Summary:

Overall, excellent work! The effort is especially commendable given the challenges at the USEPA. I only have the minor comments addressed above.
