### 2025-07-18 e37500b
Updated and renumbered vignettes.
