+++ vignettes/02_details.Rmd
@@ -1,0 +1,19 @@
+---
+title: "(2) MCSimMod Details"
+author: "Dustin Kapraun"
+date: "`r Sys.Date()`"
+output: html_vignette
+vignette: >
+  %\VignetteIndexEntry{(2) MCSimMod Details}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+```{r setup, include = FALSE}
+knitr::opts_chunk$set(
+  echo = TRUE,
+  warning = FALSE,
+  collapse = TRUE,
+  comment = "#>"
+)
+```
