+++ vignettes/04_newt_cool_demo.Rmd
@@ -1,10 +1,10 @@
 ---
-title: "(3) A Model for Newton's Law of Cooling"
+title: "(4) A Model for Newton's Law of Cooling"
 author: "Dustin Kapraun"
 date: "`r Sys.Date()`"
 output: html_vignette
 vignette: >
-  %\VignetteIndexEntry{(3) A Model for Newton's Law of Cooling}
+  %\VignetteIndexEntry{(4) A Model for Newton's Law of Cooling}
   %\VignetteEngine{knitr::rmarkdown}
   %\VignetteEncoding{UTF-8}
 ---
@@ -29,9 +29,9 @@ In order to solve an initial value problem for the Newton's Law of Cooling model
 
 ## MCSim Model Specification
 
-We used the [GNU MCSim](https://www.gnu.org/software/mcsim/) model specification language to implement the Newton's Law of Cooling model. The complete MCSim model specification file for this model, `newt_cool.model`, can be found in the `extdata` subdirectory of the **MCSimMod** package.
+We used the [GNU MCSim](https://www.gnu.org/software/mcsim/) model specification language to implement the Newton's Law of Cooling model. The complete MCSim model specification file for this model, `newt_cool.model`, can be found in the `extdata` subdirectory of the **MCSimMod** package installation directory.
 
-In the model specification file, we used the text symbol `T` to represent the state variable $T$, and the text symbols `r`, `Tamb`, and `T0` to represent the parameters $r$, $T_\textrm{amb}$, and $T_0$, respectively.
+In the model specification file, the text symbol `T` is used to represent the state variable $T$, and the text symbols `r`, `Tamb`, and `T0` are used to represent the parameters $r$, $T_\textrm{amb}$, and $T_0$, respectively.
 
 ## Building the Model
 
@@ -70,7 +70,7 @@ newt_mod$updateParms(c(T0 = 95, Tamb = 22, r = 0.7))
 newt_mod$updateY0()
 ```
 
-Note that executing the command `newt_mod$updateParms(c(T0=95, Tamb=22, r=0.7))` updated the parameter values (replacing the default values that were provided in the model specification file) and that executing the command `newt_mod$updateY0()` updated the initial value of the state variable `T` using the updated value of the parameter `T0`.
+Note that executing the command `newt_mod$updateParms(c(T0=95, Tamb=22, r=0.7))` updates the parameter values (replacing the default values that were provided in the model specification file) and that executing the command `newt_mod$updateY0()` updates the initial value of the state variable `T` using the updated value of the parameter `T0`.
 
 Finally, we can perform a simulation that provides results for the desired output times (i.e., $t = 0, 0.1, 0.2, \ldots, 5.0$) using the following commands.
 ```{r, results = 'hide'}
