### 2025-07-18 63bfd84
Renamed vignettes.
