### 2025-07-21 0b25b21
Inlcude link to issues and contact email (will need to be filled in). Also include link to GNU General Public License (future LICENSE file).
