### 2025-07-21 75f56ec
Include GPL 3.0 license and appropriate link in README
