### 2025-07-21 10fb1e0
Added new vignette 'details.Rmd'.
