+++ vignettes/details.Rmd
@@ -17,3 +17,63 @@ knitr::opts_chunk$set(
   comment = "#>"
 )
 ```
+
+The **MCSimMod** package facilitates ordinary differential equation (ODE) modeling within the R environment. **MCSimMod** allows one to solve initial value problems (IVPs) for ODE models described in the [MCSim](https://www.gnu.org/software/mcsim/) model specification language using the R package **deSolve**. This system enables users to take advantage of the flexibility and post hoc data analysis capabilities of the interpreted language R while also achieving computational speeds typically associated with compiled programming languages like C and FORTRAN. Furthermore, this system encourages modelers to use separate files for defining models and executing model simulations, which can, in turn, improve modularity and reusability of source code developed for modeling analyses.
+
+## Object-Oriented Design
+
+**MCSimMod** was designed using the object-oriented programming paradigm, in which computer programs are based on objects that comprise attributes (i.e., data) and methods (i.e., functionality). In particular, the **MCSimMod** package defines a class called `Model` that provides a template for model objects. One creates a `Model` object (i.e., an instance of the `Model` class) to represent a given ODE model. As illustrated in Figure 1, a `Model` object has both attributes (i.e., things the object "knows" about itself) and methods (i.e., things the object can "do"). Model attributes include: the name of the model (`mName`); a vector of parameter names and values (`parms`); and a vector of initial conditions (`Y0`). Model methods include functions for: translating, compiling, and loading the model (`loadModel()`); updating parameter values (`updateParms()`); updating initial conditions (`updateY0()`); and running model simulations (`runModel()`). So, for example, if `mod` is a `Model` object, it will have an attribute called `parms` that can be accessed using the R expression `mod$parms`. Similarly, `mod` will have a method called `updateParms()` that can be accessed using the R expression `mod$updateParms()`.
+
+```{r, echo = FALSE, fig.cap = "Figure 1: Model object schema", out.width = '50%'}
+knitr::include_graphics("ModelObjectFigure.png")
+```
+
+## Model Specification
+
+A **Model** can be defined using either a text file or a text string written in the MCSim model specification language. Figure 2 shows an example of an MCSim model specification file and indicates its key components: (1) names of state variables; (2) names of other "output" variables (for which values are recalculated throughout a simulation); (3) names of "input" variables (which have values that may vary throughout a simulation); (4) names and default values of parameters (which are constant throughout a simulation); (5) instructions for initializing values of parameters and state variables; (6) instructions for calculating "dynamic" aspects of the model, including values of output variables and values of derivatives for each state variable (which are expressed as ODEs); and (7) the mandatory ending keyword. (More details about the structure of the model specification text are provided in a separate vignette: "Model Specification".)
+
+```{r, echo = FALSE, fig.cap = "Figure 2: An example of an MCSim model specification", out.width = '50%'}
+knitr::include_graphics("ModelSpecificationComponents.png")
+```
+
+## Creating a Model object
+
+One can create a new **Model** object using the function `createModel()`. It is necessary to supply (as an argument to the `createModel()` function) either a string or the name of a file that contains the model specification text. For example, if `mod_string` is a character string containing model specification text, one could use the following command to create a `Model` object from that string.
+
+```{verbatim}
+model <- createModel(mString = mod_string)
+```
+
+Alternatively, one could save the model specification text in a file called `ex.model` and then use the following command to create a `Model` object from the file (assuming that the file is in the current working directory for the R session).
+
+```{verbatim}
+model <- createModel(mName = "ex")
+```
+
+Note that the suffix (or file name extension) `.model` should be included in the name of file, but it should not be included in the `mName` character string argument that is passed to the function `createModel()`. One can also provide a full file system path for the argument mName (again excluding the suffix `.model`) as shown in the following example.
[…]
