### 2025-07-22 c4da53b
Updated vignettes.
