+++ vignettes/details.Rmd
@@ -18,62 +18,95 @@ knitr::opts_chunk$set(
 )
 ```
 
-The **MCSimMod** package facilitates ordinary differential equation (ODE) modeling within the R environment. **MCSimMod** allows one to solve initial value problems (IVPs) for ODE models described in the [MCSim](https://www.gnu.org/software/mcsim/) model specification language using the R package **deSolve**. This system enables users to take advantage of the flexibility and post hoc data analysis capabilities of the interpreted language R while also achieving computational speeds typically associated with compiled programming languages like C and FORTRAN. Furthermore, this system encourages modelers to use separate files for defining models and executing model simulations, which can, in turn, improve modularity and reusability of source code developed for modeling analyses.
+Many physical and biological phenomena can be described using mathematical models based on ordinary differential equations (ODEs). In such a model, an ODE describes how a "state variable" changes (quantitatively) with respect to an independent variable (e.g., time or position). In general, an ODE model can include several state variables, each with its own ODE, so the model can be expressed as a system of ODEs. Thus, if $y$ is a vector of $n$ state variables,
+an ODE model that describes the state of the system at $t$ (i.e., at a specific time or value of the independent variable) can be expressed as
+\begin{equation}
+  \frac{\textrm{d}}{\textrm{d}t}y(t) = f(y(t), \theta, t),
+\end{equation}
+where $f$ is a vector-value function (of dimension $n$) and \theta is a vector of parameters (i.e., constants or variables other than state variables and the independent variable).
+
+In order to obtain a specific solution for a system of ODEs, one must know the initial state of the system,
+\begin{equation}
+  y_0 = y(t_0),
+\end{equation}
+where $t_0$ is the inital value of the independent variable. This last equation is often described as a statement of the "initial conditions" of the system, and solving a system of ODEs subject to such initial conditions is called solving an initial value problem (IVP).
+
+The **MCSimMod** package allows one to solve IVPs for ODE models described in the [MCSim](https://www.gnu.org/software/mcsim/) model specification language. Using **MCSimMod**, one can translate MCSim model specification text to C source code and then "compile" that code for use in R. This system enables users to take advantage of the flexibility and post hoc data analysis capabilities of the interpreted language R while also achieving computational speeds typically associated with compiled programming languages (like C and FORTRAN). Furthermore, this system encourages modelers to use separate files for defining models and executing model simulations, which can, in turn, improve modularity and reusability of source code developed for modeling analyses.
 
 ## Object-Oriented Design
 
 **MCSimMod** was designed using the object-oriented programming paradigm, in which computer programs are based on objects that comprise attributes (i.e., data) and methods (i.e., functionality). In particular, the **MCSimMod** package defines a class called `Model` that provides a template for model objects. One creates a `Model` object (i.e., an instance of the `Model` class) to represent a given ODE model. As illustrated in Figure 1, a `Model` object has both attributes (i.e., things the object "knows" about itself) and methods (i.e., things the object can "do"). Model attributes include: the name of the model (`mName`); a vector of parameter names and values (`parms`); and a vector of initial conditions (`Y0`). Model methods include functions for: translating, compiling, and loading the model (`loadModel()`); updating parameter values (`updateParms()`); updating initial conditions (`updateY0()`); and running model simulations (`runModel()`). So, for example, if `mod` is a `Model` object, it will have an attribute called `parms` that can be accessed using the R expression `mod$parms`. Similarly, `mod` will have a method called `updateParms()` that can be accessed using the R expression `mod$updateParms()`.
 
-```{r, echo = FALSE, fig.cap = "Figure 1: Model object schema", out.width = '50%'}
[…]
