### 2025-07-23 42520f6
Added vignette 'pbpk_demo.Rmd' to provide an example involving a PBPK model.
