### 2025-07-23 87680c1
Updated '_pkgdown.yml' to include the new vignette 'pbpk_demo.Rmd'.
