+++ _pkgdown.yml
@@ -14,3 +14,4 @@ articles:
   - pk1_demo
   - inputs_demo
   - events_demo
+  - pbpk_demo
