### 2025-07-23 #8492 @vipinagrawal25
# Review of MCSimMod for JOSS Submission

Hello,

Thank you to Danielskatz for the opportunity to review the MCSimMod package. Congratulations to the authors for developing MCSimMod, an R package that integrates GNU MCSim models with R for advanced PBPK modeling, complementing existing tools in the field. This package effectively bridges MCSim and R, enabling users to leverage both platforms’ strengths. Below are my comments. Once they are addressed I **recommend this paper for publication.**

## General Checks

**License**: Can you please include a LICENSE file with the contents of the GPL-2.0 (or later) license.

## Installation and Setup

The installation process is straightforward but could be improved:

- Running `devtools::document()` and `covr::report()` resulted in errors due to missing dependencies: `stringi`, `DT`, and `htmltools`. Can you please explicitly list all necessary dependencies in the documentation and include installation instructions, e.g., `install.packages(c("stringi", "DT", "htmltools"))`, to streamline setup for users.
- Consider adding a brief note with a link to the RStudio download page[](https://posit.co/download/rstudio-desktop/) for users new to R, though this is not essential.

## Documentation

- **Model Creation**: The package manual[](https://cloud.r-project.org/web/packages/MCSimMod/MCSimMod.pdf) documents `createModel` and `compileModel`, but these are not referenced in the main documentation (e.g., README or vignettes). Can you please include step-by-step examples in the primary documentation to guide users through model development.
- **Example Usage**:  Please provide a specific example demonstrating PBPK modeling with Monte Carlo or Bayesian analysis to help users apply MCSimMod to real-world problems in environmental toxicology or pharmacology.
- **Community Guidelines**: It would be nice for the community if you can include clear instructions for reporting issues (e.g., via GitHub Issues at https://github.com/USEPA/MCSimMod/issues) and seeking support (e.g., a contact email or mailing list).
- For the broader audience, can you mention potential applications of GNU MCSim models beyond PBPK, such as ecological or biochemical modeling.

## Software Paper

Overall the paper is written in very clear language. However it can be improved by addressing the following comments:

### Summary

- Explain why researchers might choose MCSimMod over existing packages like `mrgsolve`, `rxode2`, or `diffeqr`. Highlight MCSimMod’s unique features, particularly its integration with GNU MCSim for PBPK modeling, and clarify any specific functionalities not available in other packages.
- Clearly mention MCSimMod’s new or complementary contributions to the broader modeling community beyond existing tools.
- Briefly explain PBPK modeling for non-specialists.

### Statement of Need
- Consider providing a runtime comparison or implementation details for the justification for the claim that MCSimMod’s time complexity is similar to GNU MCSim
- Lines 77 and 78 contain a contradictory statement about MCSimMod being used “for only IVP” and “it is therefore.” Please revise or remove this for clarity.

Addressing these points will strengthen the clarity and impact of MCSimMod, ensuring it is recognized as a valuable contribution to the open-source software community.
