### 2025-07-28 5e29414
Remove one section from 'pbpk_demo.Rmd'.
