+++ vignettes/pbpk_demo.Rmd
@@ -140,81 +140,3 @@ Also, in addition to the time-varying quantities previously mentioned, the model
 that represents the total amount of substance that has been "lost" vs. time. This amount is calculated as [total (cumulative) amount that has entered in the organism] - [total amount currently in the organism] - [total (cumulative) amount that has been eliminated from the organism]. Therefore, if the model "conserves mass" as it should, the value of $A_\textrm{tot}(t)$ should be zero (or very close to zero for numerical solutions to an IVP) at all times $t$.
 
 Note that the model specification file uses text symbols to represent the state variables, output variables, and parameters in the model. For example, the state variable $A_\textrm{L}$ is represented by the text symbol `A_L`, the output variable $C_\textrm{L}$ is represented by the text symbol `C_L`, and the parameter $Q_\textrm{LC}$ is represented by the text symbol `Q_LC`.
-
-## Building the Model
-
-First, we load the **MCSimMod** package as follows.
-```{r, results = 'hide', message = FALSE}
-library(MCSimMod)
-```
-
-Using the following commands, we create a model object (i.e., an instance of the `Model` class) using the model specification file `pbpk_simple.model` that is included in the **MCSimMod** package.
-```{r, results = 'hide'}
-# Get the full name of the package directory that contains the example MCSim
-# model specification file.
-mod_path <- file.path(system.file(package = "MCSimMod"), "extdata")
-
-# Create a model object using the example MCSim model specification file
-# "pbpk_simple.model" included in the MCSimMod package.
-pbpk_mod_name <- file.path(mod_path, "pbpk_simple")
-pbpk_mod <- createModel(pbpk_mod_name)
-```
-
-This last command creates a `Model` object called `pbpk_mod`. Once that object is created, we can "load" the model (so that it's ready for use in a given R session) as follows.
-```{r, results = 'hide', message = FALSE}
-# Load the model.
-pbpk_mod$loadModel()
-```
-
-## Verifying Conservation of Mass
-
-We can demonstrate that the PBPK model for Substance X conserves mass by performing a simulation in which a rat is orally administered 1 mg/kg/d of Substance X and plotting total amount of substance lost, $A_\textrm{tot}(t)$, vs. time, $t$.
-
-The default values of the anatomical and physiological model parameters that are provided in the model specification file, `pbpk_simple.model`, are for a rat. This can be verified by examining the attribute `parms` of the `Model` object `pbpk_mod` as follows.
-
-```{r}
-pbpk_mod$parms
-```
-
-We want to perform a simulation in which a rat is orally administered 1 mg/kg/d of Substance X, so we will change the value of the parameter $R_\textrm{oral}$ from 0 to 1.
-
-```{r}
-pbpk_mod$updateParms(c(D_oral = 1))
-```
-
-Now, if we examine the values of the model parameters, we can see that the body mass normalized oral administration rate (represented by $D_\textrm{oral}$ and `D_oral`) has been updated from 0 to 1 mg/kg/d and the actual oral administration rate (represented by (represented by $R_\textrm{oral}$ and `R_oral`), which is a calculated parameter, has also been updated (from 0 to approximately 0.0104 mg/h).
-
-```{r}
-pbpk_mod$parms
-```
-
-We can examine the initial conditions for the state variables in the model using the following command.
-
-```{r}
-pbpk_mod$Y0
-```
-
-As stated in the model specification file, `pbpk_simple.model`, the default values for the initial conditions are all zero.
-
[…]
