### 2025-07-28 66dd00d
Removed documentation URL from DESCRIPTION file.
