### 2025-07-28 0aa5dff
Removed line with 'loadModel()' from 'pbpk_demo.Rmd'.
