+++ vignettes/pbpk_demo.Rmd
@@ -163,5 +163,5 @@ pbpk_mod <- createModel(pbpk_mod_name)
 This last command creates a `Model` object called `pbpk_mod`. Once that object is created, we can "load" the model (so that it's ready for use in a given R session) as follows.
 ```{r, results = 'hide', message = FALSE}
 # Load the model.
-pbpk_mod$loadModel()
+# pbpk_mod$loadModel()
 ```
