### 2025-07-28 fb4119c
Modified 'compileModel.R' for more complete logging of the model translation step. Added all lines back to 'pbpk_demo.Rmd'.
