### 2025-07-28 fbe23d9
Ran styler::style_pkg().
