### 2025-07-28 08ed74d
Updated 'main.yml' to use 'windows-2025' instead of 'windows-2022'.
