### 2025-07-28 a453bfe
Added debug script to 'main.yml'.
