+++ .github/workflows/main.yml
@@ -40,6 +40,16 @@ jobs:
         run: |
           R -e "styler::style_pkg('.')"
           git diff --exit-code --compact-summary || { echo "Code formatting failed; run 'styler::style_pkg()'"; exit 1; }
+      - name: Debug Issue with PBPK Vignette
+        shell: Rscript {0}
+        run: |
+          devtools::build()
+          devtools::install()
+          library(MCSimMod)
+          mod_path <- file.path(system.file(package = "MCSimMod"), "extdata")
+          pbpk_mod_name <- file.path(mod_path, "pbpk_simple")
+          pbpk_mod <- createModel(pbpk_mod_name)
+          pbpk_mod$loadModel()
       - name: Check Package
         uses: r-lib/actions/check-r-package@v2
       - name: Test Coverage
