### 2025-07-28 f694060
Simplified 'pbpk_simple.model' for debugging/testing.
