+++ .github/workflows/main.yml
@@ -40,11 +40,6 @@ jobs:
         run: |
           R -e "styler::style_pkg('.')"
           git diff --exit-code --compact-summary || { echo "Code formatting failed; run 'styler::style_pkg()'"; exit 1; }
-      - name: Debug Issue with PBPK Vignette
-        shell: Rscript {0}
-        run: |
-          devtools::document()
-          devtools::build()
       - name: Check Package
         uses: r-lib/actions/check-r-package@v2
       - name: Test Coverage
