### 2025-07-29 72ce6bd
Modified 'main.yml' so that continuous integration checks are only performed on Unix (not on Windows).
