+++ vignettes/pbpk_demo.Rmd
@@ -1,0 +1,263 @@
+---
+title: "(9) A Physiologically Based Pharmacokinetic Model"
+author: "Dustin Kapraun"
+date: "`r Sys.Date()`"
+output: html_vignette
+vignette: >
+  %\VignetteIndexEntry{(9) A Physiologically Based Pharmacokinetic Model}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+```{r setup, include = FALSE}
+knitr::opts_chunk$set(
+  echo = TRUE,
+  warning = FALSE,
+  collapse = TRUE,
+  comment = "#>"
+)
+```
+
+## Model Description
+
+Like any pharmacokinetic (PK) model, a physiologically based pharmacokinetic (PBPK) model is a mathematical description of the absorption, distribution, metabolism, and excretion of a substance by an organism. PBPK models differ from classical PK models in that many of their parameters describe actual anatomical and physiological properties of the organism, such as volumes of and blood flow rates to various organs and tissues. Figure 1 shows a schematic representation of a relatively simple PBPK model for "Substance X."
+```{r, echo = FALSE, fig.cap = "Figure 1: A simple PBPK model", out.width = "50%"}
+knitr::include_graphics("pbpk_basic_model_schematic.png")
+```
+Substance X can enter the body through oral and/or intravenous routes and it is eliminated via saturable metabolism in the liver. The kinetics of distribution are assumbed to be "perfusion limited," so concentrations in different tissues depend on steady state concentration ratios, or "partition coefficients," for each type of tissue.
+
+The state variables for the PBPK model include:
+
+* $A_\textrm{G}$, the amount (of Substance X) in the gut lumen (mg);
+* $A_\textrm{L}$, the amount in the liver (mg);
+* $A_\textrm{F}$, the amount in the fat (mg);
+* $A_\textrm{AB}$, the amount in the arterial blood (mg);
+* $A_\textrm{VB}$, the amount in the venous blood (mg);
+* $A_\textrm{R}$, the amount in the rest of body (mg);
+* $A_\textrm{dose}$, the cumulative amount that has been dosed (mg); and
+* $A_\textrm{m}$, the cumulative amount that has been metabolized (mg).
+
+The anatomical and physiological parameters are:
+
+* $M$, the body mass of the organism (kg);
+* $Q_\textrm{CC}$, the cardiac output constant (L/h/kg${}^{0.75}$, i.e., liters per hour per kg of body mass to the 3/4 power);
+* $Q_\textrm{LC}$, the constant fractional blood flow to the liver;
+* $Q_\textrm{FC}$, the constant fractional blood flow to the fat;
+* $V_\textrm{LC}$, the constant volume fraction of body mass that is the liver (L/kg);
+* $V_\textrm{FC}$, the constant volume fraction of body mass that is the fat (L/kg);
+* $V_\textrm{ABC}$, the constant volume fraction of body mass that is the arterial blood (L/kg); and
+* $V_\textrm{VBC}$, the constant volume fraction of body mass that is the venous blood (L/kg).
+
+Values for anatomical and physiological parameters for rats and humans are shown in the following table.
+
+```{r table1, echo=FALSE, message=FALSE, warnings=FALSE, results='asis'}
+tabl <- "
+| Parameter         | Units             | Rat Value | Human Value |
[…]
