+++ _pkgdown.yml
@@ -13,3 +13,4 @@ articles:
   - pk1_demo
   - inputs_demo
   - events_demo
+  - pbpk_demo
