### 2025-07-29 feda692
Updated 'paper.md' to explain why users might choose MCSimMod over existing R packages for solving ODE IVPs.
