+++ paper/paper.bib
@@ -35,6 +35,18 @@
   doi = {10.32614/cran.package.mrgsolve},
 }
 
+@article{diffeqr,
+  doi = {10.5334/jors.151},
+  title = {DifferentialEquations.jl – A Performant and Feature-Rich Ecosystem for Solving Differential Equations in Julia},
+  author = {Chris Rackauckas and Qing Nie},
+  journal = {The Journal of Open Source Software},
+  volume = {5},
+  number = {1},
+  year = {2017},
+  url = {https://openresearchsoftware.metajnl.com/articles/10.5334/jors.151/},
+  note = {R package version 2.1.0},
+  }
+
 @article{MCSim,
   title = {GNU MCSim: Bayesian Statistical Inference for SBML-Coded Systems Biology Models},
   author = {Frederic Y. Bois},
