### 2025-07-29 c22225c
build vignettes
