### 2025-07-30 787c7fc
Edited 'pbpk_demo.Rmd' so that R code chunks are not evaluated. Tranlsating and compiling 'pbpk_simple.model' seems to require so much time that it causes devtools::build() to time out.
