+++ vignettes/pbpk_demo.Rmd
@@ -144,12 +144,12 @@ Note that the model specification file uses text symbols to represent the state 
 ## Building the Model
 
 First, we load the **MCSimMod** package as follows.
-```{r, results = 'hide', message = FALSE}
+```{r, eval = FALSE}
 library(MCSimMod)
 ```
 
 Using the following commands, we create a model object (i.e., an instance of the `Model` class) using the model specification file `pbpk_simple.model` that is included in the **MCSimMod** package.
-```{r, results = 'hide'}
+```{r, eval = FALSE}
 # Get the full name of the package directory that contains the example MCSim
 # model specification file.
 mod_path <- file.path(system.file(package = "MCSimMod"), "extdata")
@@ -161,7 +161,7 @@ pbpk_mod <- createModel(pbpk_mod_name)
 ```
 
 This last command creates a `Model` object called `pbpk_mod`. Once that object is created, we can "load" the model (so that it's ready for use in a given R session) as follows.
-```{r, results = 'hide', message = FALSE}
+```{r, eval = FALSE}
 # Load the model.
 pbpk_mod$loadModel()
 ```
@@ -172,45 +172,67 @@ We can demonstrate that the PBPK model for Substance X conserves mass by perform
 
 The default values of the anatomical and physiological model parameters that are provided in the model specification file, `pbpk_simple.model`, are for a rat, and these are the values that are assigned to the parameters when one calls the `loadModel()` method. This can be verified by examining the attribute `parms` of the `Model` object `pbpk_mod` as follows.
 
-```{r}
+```{r, eval = FALSE}
 pbpk_mod$parms
+#>          M       Q_CC       Q_LC       Q_FC       V_LC       V_FC      V_ABC 
+#>  0.2500000 15.0000000  0.2100000  0.0600000  0.0400000  0.0700000  0.0200000 
+#>      V_VBC        P_L        P_F        P_R     V_maxC        K_m      k_abs 
+#>  0.0500000  3.5000000 86.5000000  2.3000000 10.0000000  6.0000000  1.2500000 
+#>     D_oral       D_IV        Q_C        Q_L        Q_F       Q_RC        Q_R 
+#>  0.0000000  0.0000000  5.3033009  1.1136932  0.3181981  0.7300000  3.8714096 
+#>        V_L        V_F       V_AB       V_VB       V_RC        V_R      V_max 
+#>  0.0100000  0.0175000  0.0050000  0.0125000  0.8200000  0.2050000  3.5355339 
+#>     R_oral       R_IV 
+#>  0.0000000  0.0000000
 ```
 
 We want to perform a simulation in which a rat is orally administered 1 mg/kg/d of Substance X, so we will change the value of the parameter $D_\textrm{oral}$ from 0 to 1.
 
-```{r}
+```{r, eval = FALSE}
 pbpk_mod$updateParms(c(D_oral = 1))
 ```
 
 Now, if we examine the values of the model parameters, we can see that the body mass normalized oral administration rate (represented by $D_\textrm{oral}$ and `D_oral`) has been updated from 0 to 1 mg/kg/d and the actual oral administration rate (represented by (represented by $R_\textrm{oral}$ and `R_oral`), which is a calculated parameter, has also been updated (from 0 to approximately 0.0104 mg/h).
 
-```{r}
+```{r, eval = FALSE}
 pbpk_mod$parms
+#>           M        Q_CC        Q_LC        Q_FC        V_LC        V_FC 
+#>  0.25000000 15.00000000  0.21000000  0.06000000  0.04000000  0.07000000 
[…]
