+++ vignettes/pbpk_demo.Rmd
@@ -174,15 +174,15 @@ The default values of the anatomical and physiological model parameters that are
 
 ```{r, eval = FALSE}
 pbpk_mod$parms
-#>          M       Q_CC       Q_LC       Q_FC       V_LC       V_FC      V_ABC 
-#>  0.2500000 15.0000000  0.2100000  0.0600000  0.0400000  0.0700000  0.0200000 
-#>      V_VBC        P_L        P_F        P_R     V_maxC        K_m      k_abs 
-#>  0.0500000  3.5000000 86.5000000  2.3000000 10.0000000  6.0000000  1.2500000 
-#>     D_oral       D_IV        Q_C        Q_L        Q_F       Q_RC        Q_R 
-#>  0.0000000  0.0000000  5.3033009  1.1136932  0.3181981  0.7300000  3.8714096 
-#>        V_L        V_F       V_AB       V_VB       V_RC        V_R      V_max 
-#>  0.0100000  0.0175000  0.0050000  0.0125000  0.8200000  0.2050000  3.5355339 
-#>     R_oral       R_IV 
+#>          M       Q_CC       Q_LC       Q_FC       V_LC       V_FC      V_ABC
+#>  0.2500000 15.0000000  0.2100000  0.0600000  0.0400000  0.0700000  0.0200000
+#>      V_VBC        P_L        P_F        P_R     V_maxC        K_m      k_abs
+#>  0.0500000  3.5000000 86.5000000  2.3000000 10.0000000  6.0000000  1.2500000
+#>     D_oral       D_IV        Q_C        Q_L        Q_F       Q_RC        Q_R
+#>  0.0000000  0.0000000  5.3033009  1.1136932  0.3181981  0.7300000  3.8714096
+#>        V_L        V_F       V_AB       V_VB       V_RC        V_R      V_max
+#>  0.0100000  0.0175000  0.0050000  0.0125000  0.8200000  0.2050000  3.5355339
+#>     R_oral       R_IV
 #>  0.0000000  0.0000000
 ```
 
@@ -196,15 +196,15 @@ Now, if we examine the values of the model parameters, we can see that the body 
 
 ```{r, eval = FALSE}
 pbpk_mod$parms
-#>           M        Q_CC        Q_LC        Q_FC        V_LC        V_FC 
-#>  0.25000000 15.00000000  0.21000000  0.06000000  0.04000000  0.07000000 
-#>       V_ABC       V_VBC         P_L         P_F         P_R      V_maxC 
-#>  0.02000000  0.05000000  3.50000000 86.50000000  2.30000000 10.00000000 
-#>         K_m       k_abs      D_oral        D_IV         Q_C         Q_L 
-#>  6.00000000  1.25000000  1.00000000  0.00000000  5.30330086  1.11369318 
-#>         Q_F        Q_RC         Q_R         V_L         V_F        V_AB 
-#>  0.31819805  0.73000000  3.87140963  0.01000000  0.01750000  0.00500000 
-#>        V_VB        V_RC         V_R       V_max      R_oral        R_IV 
+#>           M        Q_CC        Q_LC        Q_FC        V_LC        V_FC
+#>  0.25000000 15.00000000  0.21000000  0.06000000  0.04000000  0.07000000
+#>       V_ABC       V_VBC         P_L         P_F         P_R      V_maxC
+#>  0.02000000  0.05000000  3.50000000 86.50000000  2.30000000 10.00000000
+#>         K_m       k_abs      D_oral        D_IV         Q_C         Q_L
+#>  6.00000000  1.25000000  1.00000000  0.00000000  5.30330086  1.11369318
[…]
