### 2025-07-30 f139cde
Add R plot figures to vignette 'pbpk_demo.Rmd'.
