### 2025-07-30 6d322a4
Updated 'paper.md' and 'README.md' with new text.
