+++ paper/paper.bib
@@ -35,18 +35,6 @@
   doi = {10.32614/cran.package.mrgsolve},
 }
 
-@article{diffeqr,
-  doi = {10.5334/jors.151},
-  title = {DifferentialEquations.jl – A Performant and Feature-Rich Ecosystem for Solving Differential Equations in Julia},
-  author = {Chris Rackauckas and Qing Nie},
-  journal = {The Journal of Open Source Software},
-  volume = {5},
-  number = {1},
-  year = {2017},
-  url = {https://openresearchsoftware.metajnl.com/articles/10.5334/jors.151/},
-  note = {R package version 2.1.0},
-  }
-
 @article{MCSim,
   title = {GNU MCSim: Bayesian Statistical Inference for SBML-Coded Systems Biology Models},
   author = {Frederic Y. Bois},
@@ -57,12 +45,15 @@
   doi = {10.1093/bioinformatics/btp162},
 }
 
-@misc{MCSim_webpage,
-  author = {Frederic Y. Bois},
-  year = {2024},
-  title = {GNU MCSim},
-  howpublished ={\url{https://www.gnu.org/software/mcsim/}},
-  note = {Accessed: 2025-01-23},
+@article{FME,
+  title = {Inverse Modelling, Sensitivity and Monte Carlo Analysis in {R} Using Package {FME}},
+  author = {Karline Soetaert and Thomas Petzoldt},
+  journal = {Journal of Statistical Software},
+  year = {2010},
+  volume = {33},
+  number = {3},
+  pages = {1--28},
+  doi = {10.18637/jss.v033.i03},
 }
 
 @article{Pearce2017_httk,
@@ -73,19 +64,7 @@
    journal={Journal of Statistical Software},
    author={Pearce, Robert G. and Setzer, R. Woodrow and Strope, Cory L. and Sipes, Nisha S. and Wambaugh, John F.},
    year={2017},
-   pages={1–-26}
-}
-
-@article{Kapraun2020_naphthalene,
-  author = {Kapraun, Dustin F and Schlosser, Paul M and Nylander-French, Leena A and Kim, David and Yost, Erin E and Druwe, Ingrid L},
-  title = {A Physiologically Based Pharmacokinetic Model for Naphthalene With Inhalation and Skin Routes of Exposure},
-  journal = {Toxicological Sciences},
-  volume = {177},
-  number = {2},
-  pages = {377--391},
-  year = {2020},
-  month = {07},
-  doi = {10.1093/toxsci/kfaa117},
+   pages={1--26}
 }
 
 @article{Clewell2019_chloroprene,
@@ -147,5 +126,16 @@
   publisher={CRC Press},
   address={Boca Raton, Florida},
 }
-  
-  
+
+@article{diffeqr,
+  doi = {10.5334/jors.151},
+  title = {DifferentialEquations.jl ??? A Performant and Feature-Rich Ecosystem for Solving Differential Equations in Julia},
+  doi = {10.5334/jors.151},
+  journal = {The Journal of Open Source Software},
+  author = {Chris Rackauckas and Qing Nie},
+  year = {2017},
+  volume = {5},
+  number = {1},
+  url = {https://openresearchsoftware.metajnl.com/articles/10.5334/jors.151/},
+  note = {R package version 2.1.0},
+}
