+++ paper/paper.md
@@ -25,7 +25,7 @@ affiliations:
    index: 1
  - name: Pacific Northwest National Laboratory, U.S.A.
    index: 2
-date: 29 July 2025
+date: 30 July 2025
 bibliography: paper.bib
 ---
 
@@ -45,11 +45,13 @@ where $t_0$ is the inital value of the independent variable. \autoref{eq:ic_gene
 
 For the R programming language and environment [@R], there are at least two packages available that facilitate solving of IVPs. The R package `deSolve` [@deSolve] can be used to solve IVPs for ODEs that have been encoded either in R or in a compiled language, such as C or Fortran. For models encoded in a compiled language, one can compile the model source code to generate machine code that typically executes much more quickly than R code, which must be "interpreted" anew each time it's executed on a computer. The `deSolve` package includes functions that provide interfaces to well-documented, public-domain IVP solution routines implemented in FORTRAN, including four ODE integrators from the package ODEPACK, and in C, including solvers of the Runge-Kutta family. The R package `mrgsolve` [@mrgsolve] also includes functions that can be used to solve IVPs. These `mrgsolve` functions provide interfaces to IVP solution routines implemented in C++. Despite their different implementations, the packages `deSolve` and `mrgsolve` use many of the same underlying IVP solution algorithms.
 
-We developed the R package [`MCSimMod`](https://cran.r-project.org/package=MCSimMod) to facilitate ODE modeling within the R environment. `MCSimMod` allows one to solve IVPs for ODE models that have been described in the MCSim [@MCSim] model specification language using ODE integration functions from the R package `deSolve` [@deSolve]. This system enables users to take advantage of the flexibility and post hoc data analysis capabilities of the interpreted language R while also achieving computational speeds typically associated with compiled programming languages like C and FORTRAN. Furthermore, this system encourages modelers to use separate files for defining models and executing model simulations, which can, in turn, improve modularity and reusability of source code developed for modeling analyses. While other packages, including `mrgsolve` [@mrgsolve] and `diffeqr` [@diffeqr], facilitate ODE modeling in the R environment, `MCSimMod` is unique in that it allows users to work directly with ODE models encoded in the MCSim model specification language.
+We developed the R package [`MCSimMod`](https://cran.r-project.org/package=MCSimMod) to facilitate ODE modeling within the R environment. `MCSimMod` allows one to solve IVPs for ODE models that have been described in the MCSim [@MCSim] model specification language using ODE integration functions from the R package `deSolve` [@deSolve]. This system enables users to take advantage of the flexibility and post hoc data analysis capabilities of the interpreted language R while also achieving computational speeds typically associated with compiled programming languages like C and FORTRAN. Furthermore, this system encourages modelers to use separate files for defining models and executing model simulations, which can, in turn, improve modularity and reusability of source code developed for modeling analyses. Examples of model specification files and vignettes that demonstrate how to use them to perform simulations are included in the `MCSimMod` package. While other packages, including `mrgsolve` [@mrgsolve] and `diffeqr` [@diffeqr], facilitate ODE modeling in the R environment, `MCSimMod` is unique in that it allows users to work directly with ODE models encoded in the MCSim model specification language.
+
[…]
