+++ README.md
@@ -7,11 +7,9 @@
 [![GitHub Actions Workflow Status](https://img.shields.io/github/actions/workflow/status/USEPA/MCSimMod/main.yml)](https://github.com/USEPA/MCSimMod/actions?query=branch%3Amain)
 [![GitHub Release](https://img.shields.io/github/v/release/USEPA/MCSimMod)](https://github.com/USEPA/MCSimMod/releases)
 
-MCSimMod is an R package that facilitates ordinary differential equation (ODE) modeling. It allows one to perform simulations for ODE models that are encoded in the [GNU MCSim](https://www.gnu.org/software/mcsim/) model specification language using ODE solvers from the R package [deSolve](https://cran.r-project.org/web/packages/deSolve/index.html).
+`MCSimMod` is an R package that facilitates ordinary differential equation (ODE) modeling. Many physical and biological phenomena can be described using mathematical models based on ODEs, and several example ODE models are described in "vignettes" included with the `MCSimMod` package. `MCSimMod` allows one to perform simulations for ODE models that are encoded in the [GNU MCSim](https://www.gnu.org/software/mcsim/) model specification language using ODE solvers from the R package [deSolve](https://cran.r-project.org/web/packages/deSolve/index.html).
 
 Documentation is available at https://usepa.github.io/MCSimMod.
-
-For users new to R, we recommend [RStudio](https://posit.co/download/rstudio-desktop/) to install packages and edit/run scripts.
 
 To work with `MCSimMod`, the package `deSolve` must be installed.
 ```R
@@ -22,7 +20,7 @@ install.packages("deSolve")
 
 As a prerequisite for installation on a Windows operating system, be sure to install [RTools](https://cran.r-project.org/bin/windows/Rtools/). RTools includes utilities that compile C source code for use in R. Installation of RTools is not required when using `MCSimMod` on a Unix operating system.
 
-To install the latest stable version from [CRAN](https://cran.r-project.org/web/packages/MCSimMod/), use the command:
+To install the latest stable version of `MCSimMod` from [CRAN](https://cran.r-project.org/web/packages/MCSimMod/), use the command:
 
 ```R
 install.packages("MCSimMod")
@@ -45,12 +43,8 @@ To learn about the `MCSimMod` package and how to use it, check out the vignettes
 browseVignettes(package = "MCSimMod")
 ```
 
-If you are interested in contributing or want to report a bug, please start a discussion or submit an issue [here](https://github.com/USEPA/MCSimMod.git).
-
 ## Contact
-To report an issue with MCSimMod, please use the [issue tracker](https://github.com/USEPA/MCSimMod/issues).
-
-For any technical questions regarding modeling and usage of MCSimMod, [send us an e-mail](mailto:some.one@epa.gov).
+To report an issue with `MCSimMod`, please use the [issue tracker](https://github.com/USEPA/MCSimMod/issues).
 
 ## License
 [GNU General Public License 3.0](https://github.com/USEPA/MCSimMod/blob/main/LICENSE)
@@ -63,7 +57,7 @@ devtools::document()
 devtools::build()
 devtools::install()
 devtools::test()
-covr::report(file='coverage_html/index.html')
+covr::report(file = "coverage_html/index.html")
 ```
 
 You can also issue these commands from a Windows or Unix command line terminal as follows.
@@ -72,20 +66,20 @@ R -e "devtools::document()"
 R -e "devtools::build()"
 R -e "devtools::install()"
 R -e "devtools::test()"
-R -e "covr::report(file='coverage_html/index.html')"
+R -e "covr::report(file = 'coverage_html/index.html')"
 ```
 
 ## Code Formatting
[…]

[…]
