### 2025-07-31 ca4212d
Updated 'paper.md' and 'paper.bib'.
