+++ paper/paper.bib
@@ -128,10 +128,9 @@
 }
 
 @article{diffeqr,
+  title = {DifferentialEquations.jl -- A Performant and Feature-Rich Ecosystem for Solving Differential Equations in Julia},
   doi = {10.5334/jors.151},
-  title = {DifferentialEquations.jl ??? A Performant and Feature-Rich Ecosystem for Solving Differential Equations in Julia},
-  doi = {10.5334/jors.151},
-  journal = {The Journal of Open Source Software},
+  journal = {The Journal of Open Research Software},
   author = {Chris Rackauckas and Qing Nie},
   year = {2017},
   volume = {5},
