### 2025-07-31 3e1cd09
Updated 'DESCRIPTION' file.
