+++ DESCRIPTION
@@ -8,7 +8,7 @@ Authors@R: c(
   person("Ryan D.", "Friese", role="aut", comment = c(ORCID = "0000-0002-4121-2195")),
   person("Frederic Y.", "Bois", role="ctb", comment = c(ORCID = " 0000-0002-4154-0391")),
   person("Free Software Foundation, Inc.", role="cph")
-)
+  )
 Description: Tools that facilitate ordinary differential equation (ODE) modeling in 'R'. This package allows one to perform simulations for ODE models that are encoded in the GNU 'MCSim' model specification language (Bois, 2009) <doi:10.1093/bioinformatics/btp162> using ODE solvers from the 'R' package 'deSolve' (Soetaert et al., 2010) <doi:10.18637/jss.v033.i09>.
 Imports:
   deSolve,
