### 2025-07-31 cde83cd
Corrected typographical error in 'details.Rmd'.
