### 2025-07-31 #8492 @danielskatz
👋 @vipinagrawal25 & @nanhung - please take a look again and see if you can now complete your checklists, and if not, tell us what else you think is needed.
