### 2025-08-01 88d3723
Corrected typographical error in 'newt_cool_demo.Rmd'.
