### 2025-08-01 4ff186c
Minor modifications to three vignettes.
