### 2025-08-04 1fb9165
Updated badges in 'README' file.
