### 2025-08-05 #8492 @danielskatz
👋 @vipinagrawal25 & @nanhung - please take a look again and see if you can now complete your checklists, and if not, tell us what else you think is needed.

Specifically, @nanhung - is there still an issue you see with the license?  (note that JOSS doesn't require a LICENSE file but just a license, and that standard practice for R packages that are in CRAN is that the license is listed in the DESCRIPTION file - see https://cran.rstudio.com. I do think that adding the license to the README would help though.)
And @vipinagrawal25 - can you make progress on the issues you haven't checked off?
