### 2025-08-15 3caf269
Updated version number to 1.0.
