+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: MCSimMod
 Title: Working with 'MCSim' Models
-Version: 0.9.1.9000
+Version: 1.0
 Authors@R: c(
   person("Dustin F.", "Kapraun", role=c("aut", "cre"), email = "kapraun.dustin@epa.gov", comment = c(ORCID = "0000-0001-5570-6383")),
   person("Todd J.", "Zurlinden", role="aut", comment = c(ORCID = "0000-0003-1372-3913")),
