+++ paper/paper.bib
@@ -100,7 +100,7 @@
   doi = {10.1093/toxsci/kfab063},
 }
 
-@book{Lotka2025,
+@book{Lotka1925,
   author={Lotka, A. J.},
   title={Elements of Physical Biology},
   year={1925},
