### 2025-08-15 6c6c037
Made minor changes in response to Copilot review.
