### 2025-08-15 671ac03
Address Review Comments (#40)
* Initial commit for 'manuscript' branch.

* Update draft-pdf.yml

* Updated 'paper.md'.

* (auto) Paper PDF Draft

* Updated 'paper.md'.

* (auto) Paper PDF Draft

* Updated 'paper.md'.

* (auto) Paper PDF Draft

* Updated 'paper.md'.

* (auto) Paper PDF Draft

* Updated 'paper.md'.

* (auto) Paper PDF Draft

* Updated 'paper.md'.

* (auto) Paper PDF Draft

* Updated  'paper.bib' with DOIs.

* (auto) Paper PDF Draft

* Add RStudio link to README

* Updated and renumbered vignettes.

* Updated _pkgdown.yml

* Updated vignette '02_details.Rmd' to conform with style
[…]
