+++ paper/paper.md
@@ -1,0 +1,64 @@
+---
+title: 'MCSimMod: An R Package for Working with Ordinary Differential Equation Models Encoded in the MCSim Model Specification Language'
+tags:
+  - R
+  - MCSim
+  - ordinary differential equations
+  - mathematical models
+  - initial value problems
+authors:
+  - name: Dustin F. Kapraun
+    orcid: 0000-0001-5570-6383
+    corresponding: true (This is how to denote the corresponding author)
+    affiliation: 1
+  - name: Todd J. Zurlinden
+    orcid: 0000-0003-1372-3913
+    affiliation: 1
+  - name: Ryan D. Friese
+    orcid: 0000-0002-4121-2195
+    affiliation: 2
+  - name: Andrew J. Shapiro
+    orcid: 0000-0002-5233-8092
+    affiliation: 1
+affiliations:
+ - name: U.S. Environmental Protection Agency, U.S.A.
+   index: 1
+ - name: Pacific Northwest National Laboratory, U.S.A.
+   index: 2
+date: 31 July 2025
+bibliography: paper.bib
+---
+
+# Summary
+
+Many physical and biological phenomena can be described using mathematical models based on ordinary differential equations (ODEs). In such a model, an ODE describes how a "state variable" changes (quantitatively) with respect to an independent variable (e.g., time or position). In general, an ODE model can include several state variables, each with its own ODE, so the model can be expressed as a system of ODEs. Thus, if $y$ is a vector of $n$ state variables, an ODE model that describes the state of the system at $t$ (i.e., at a specific time or value of the independent variable) can be expressed as
+\begin{equation} \label{eq:ode_general}
+  \frac{\textrm{d}}{\textrm{d}t} y(t) = f(y(t), \theta, t),
+\end{equation}
+where $f$ is a vector-valued function (of dimension $n$) and $\theta$ is a vector of parameters (i.e., constants or variables other than the state variables and the independent variable).
+
+In order to obtain a specific solution for a system of ODEs, one must know the initial state of the system,
+\begin{equation} \label{eq:ic_general}
+  y_0 = y(t_0),
+\end{equation}
+where $t_0$ is the inital value of the independent variable. \autoref{eq:ic_general} is often described as a statement of the "initial conditions" of the system, and solving a system of ODEs subject to such initial conditions is called solving an initial value problem (IVP).
+
+For the R programming language and environment [@R], there are at least two packages available that facilitate solving of IVPs. The R package `deSolve` [@deSolve] can be used to solve IVPs for ODEs that have been encoded either in R or in a compiled language, such as C or Fortran. For models encoded in a compiled language, one can compile the model source code to generate machine code that typically executes much more quickly than R code, which must be "interpreted" anew each time it's executed on a computer. The `deSolve` package includes functions that provide interfaces to well-documented, public-domain IVP solution routines implemented in FORTRAN, including four ODE integrators from the package ODEPACK, and in C, including solvers of the Runge-Kutta family. The R package `mrgsolve` [@mrgsolve] also includes functions that can be used to solve IVPs. These `mrgsolve` functions provide interfaces to IVP solution routines implemented in C++. Despite their different implementations, the packages `deSolve` and `mrgsolve` use many of the same underlying IVP solution algorithms.
+
[…]
