+++ paper/paper.bib
@@ -1,0 +1,140 @@
+@manual{R,
+  title = {R: A Language and Environment for Statistical Computing},
+  author = {{R Core Team}},
+  organization = {R Foundation for Statistical Computing},
+  address = {Vienna, Austria},
+  year = {2024},
+  url = {https://www.R-project.org/},
+}
+
+@article{deSolve,
+  title = {Solving Differential Equations in {R}: Package de{S}olve},
+  author = {Karline Soetaert and Thomas Petzoldt and R. Woodrow Setzer},
+  journal = {Journal of Statistical Software},
+  volume = {33},
+  number = {9},
+  pages = {1--25},
+  year = {2010},
+  doi = {10.18637/jss.v033.i09},
+}
+
+@manual{deSolve_documentation,
+  title = {deSolve: Solvers for Initial Value Problems of Differential Equations},
+  author = {Karline Soetaert and Thomas Petzoldt and R. Woodrow Setzer and Peter N. Brown and George D. Byrne and Ernst Hairer and Alan C. Hindmarsh and Cleve Moler and Linda R. Petzold and Youcef Saad and Clement W. Ulrich},
+  year = {2023},
+  note = {R package version 1.40},
+  url = {https://cran.r-project.org/web/packages/deSolve/deSolve.pdf},
+}
+
+@manual{mrgsolve,
+  title = {mrgsolve: Simulate from ODE-Based Models},
+  author = {Kyle T Baron},
+  year = {2024},
+  note = {R package version 1.5.2},
+  url = {https://CRAN.R-project.org/package=mrgsolve},
+  doi = {10.32614/cran.package.mrgsolve},
+}
+
+@article{MCSim,
+  title = {GNU MCSim: Bayesian Statistical Inference for SBML-Coded Systems Biology Models},
+  author = {Frederic Y. Bois},
+  journal = {Bioinformatics},
+  volume = {25},
+  pages = {1453--1454},
+  year = {2009},
+  doi = {10.1093/bioinformatics/btp162},
+}
+
+@article{FME,
+  title = {Inverse Modelling, Sensitivity and Monte Carlo Analysis in {R} Using Package {FME}},
+  author = {Karline Soetaert and Thomas Petzoldt},
+  journal = {Journal of Statistical Software},
+  year = {2010},
+  volume = {33},
+  number = {3},
+  pages = {1--28},
+  doi = {10.18637/jss.v033.i03},
+}
+
+@article{Pearce2017_httk,
+   title={httk: R Package for High-Throughput Toxicokinetics},
+   volume={79},
+   doi={10.18637/jss.v079.i04},
+   number={4},
+   journal={Journal of Statistical Software},
+   author={Pearce, Robert G. and Setzer, R. Woodrow and Strope, Cory L. and Sipes, Nisha S. and Wambaugh, John F.},
+   year={2017},
+   pages={1--26}
+}
+
+@article{Clewell2019_chloroprene,
+  author = {Harvey J. Clewell and Jerry L. Campbell and Cynthia Van Landingham and Allison Franzen and Miyoung Yoon and Darol E. Dodd and Melvin E. Andersen and P. Robinan Gentry},
+  title = {Incorporation of in vitro metabolism data and physiologically based pharmacokinetic modeling in a risk assessment for chloroprene},
+  journal = {Inhalation Toxicology},
+  volume = {31},
+  number = {13-14},
+  pages = {468--483},
+  year = {2019},
+  publisher = {Taylor \& Francis},
+  doi = {10.1080/08958378.2020.1715513},
+}
+
+@article{Campbell2023_chloroprene,
+  author={Campbell, J. L.  and Clewell, H. J.  and Van Landingham, C.  and Gentry, P. R.  and Andersen, M. E.},
[…]
