### 2025-08-15 #8492 @dfkapraun
Hi @danielskatz -

As you requested, I have created a tagged release of MCSimMod:
https://github.com/USEPA/MCSimMod/releases/tag/v1.0 .

I also created an archival deposit in Zenodo:
https://doi.org/10.5281/zenodo.16883665 .

I'm not sure how to make modifications to the metadata for the Zenodo
record. If you think any changes are needed, please advise!

Best regards,
Dustin

On Wed, Aug 13, 2025 at 9:31 AM Daniel S. Katz ***@***.***>
wrote:

> *danielskatz* left a comment (openjournals/joss-reviews#8492)
> <https://github.com/openjournals/joss-reviews/issues/8492#issuecomment-3183940174>
>
> 👋 @dfkapraun <https://github.com/dfkapraun> - At this point could you:
>
>    - Make a tagged release of your software, and list the version tag of
>    the archived version here.
>    - Archive the reviewed software in Zenodo or a similar service (e.g.,
>    figshare, an institutional repository)
>    - Check the archival deposit (e.g., in Zenodo) has the correct
>    metadata. This includes the title (should match the paper title) and author
>    list (make sure the list is correct and people who only made a small fix
>    are not on it). You may also add the authors' ORCID.
>    - Please list the DOI of the archived version here.
>
> I can then move forward with accepting the submission.
>
