### 2025-08-15 a630b3e
Remove workflow action that adds and commits paper PDF. (#42)
