+++ .github/workflows/draft-pdf.yml
@@ -27,9 +27,9 @@ jobs:
         # PDF. Note, this should be the same directory as the input
         # paper.md
         path: paper/paper.pdf
-    - name: Commit PDF to repository
-      uses: EndBug/add-and-commit@v9
-      with:
-        message: '(auto) Paper PDF Draft'
-        # This should be the path to the paper within your repo.
-        add: 'paper/paper.pdf'
+    # - name: Commit PDF to repository
+    #   uses: EndBug/add-and-commit@v9
+    #   with:
+    #     message: '(auto) Paper PDF Draft'
+    #     # This should be the path to the paper within your repo.
+    #     add: 'paper/paper.pdf'
