### 2025-08-15 177bd3a
Update README.md
Add Zenodo DOI for release; remove a few other badges
