### 2025-08-15 #8492 @danielskatz
> I'm not sure how to make modifications to the metadata for the Zenodo
record. If you think any changes are needed, please advise!

yes, the title should be changed to that of the paper, and you might want to add ORCIDs for the authors.  See https://help.zenodo.org/docs/deposit/manage-records/.
