### 2025-08-15 #8492 @dfkapraun
> 👋 [@dfkapraun](https://github.com/dfkapraun) - the title of the zenodo deposit isn't the same as the title of the paper. And as I mentioned, you might want to add ORCIDs for the authors there.

@danielskatz - These issues have now been corrected. Zenodo record has correct title and ORCIDs are included!
