### 2025-08-15 #8492 @dfkapraun
@danielskatz @vipinagrawal25 @nanhung - Thank you!!

-Dustin Kapraun

On Fri, Aug 15, 2025 at 5:09 PM Daniel S. Katz ***@***.***>
wrote:

> *danielskatz* left a comment (openjournals/joss-reviews#8492)
> <https://github.com/openjournals/joss-reviews/issues/8492#issuecomment-3192768074>
>
> Congratulations to @dfkapraun <https://github.com/dfkapraun> (Dustin F.
> Kapraun) and co-authors on your JOSS publication!!
>
> And thanks to @vipinagrawal25 <https://github.com/vipinagrawal25> &
> @nanhung <https://github.com/nanhung> for reviewing!
> JOSS depends on volunteers and we couldn't do this without you
>
